Move Clifford loss test to simulation suite
